# TST: Speed up repeated jax tests