Add pytest option --jax-pcc for the jax persistent compilation cache. This uses a jax config to set cache options.
Total size of cache for all tests (well, really four main ones) is only 2.7M on linux, 416 files.
On subsequent calls, reduces runtime from
4.16s call     test_optimizers.py::test_fit[SBR]
2.98s call     test_optimizers.py::test_complexity_not_fitted[SBR]
2.89s call     test_optimizers.py::test_pickle[SBR-opt_args1]
2.89s call     test_optimizers.py::test_sbr_accurate

To:
1.94s call     test_optimizers.py::test_fit[SBR]
1.56s call     test_optimizers.py::test_sbr_accurate
1.23s call     test_optimizers.py::test_pickle[SBR-opt_args1]
1.18s call     test_optimizers.py::test_complexity_not_fitted[SBR]